Move single panel qc plot into its own function so I can use it for the planes
